backend/requirements/requirementsProd.txt
argcomplete==1.9.3
boto3==1.9.23
botocore==1.12.23
certifi==2018.8.24
cfn-flip==1.0.3
chardet==3.0.4
Click==7.0
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.7
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
placebo==0.8.2
psycopg2==2.7.5
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.3
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.0
